[diffusion] Fold ComfyUI profile, session, and checkpoint spec into two modules
Keep --comfyui-mode assembly and the run cache together as comfyui_mode, and
move the single-file DiT spec plus loader into the existing checkpoints package.
